Changed all prices to Float and formatted them (hopefully) everywhere to 2 decimals

+++ src/main/java/com/ingsw/flyingdutchman/controller/ThresholdManagement.java
@@ -166,7 +166,7 @@ public class ThresholdManagement {
 
             try {
                 daoFactory.getThresholdDAO().create(
-                        Integer.parseInt(request.getParameter("price")),
+                        Float.parseFloat(request.getParameter("price")),
                         timestamp,
                         loggedUser,
                         auction
